requirements.txt
# Django
Django==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.9

# Images
Pillow>=11.3.0
sorl-thumbnail==12.10.0

# Forms and utilities
django-widget-tweaks==1.5.0

# Environment variables
python-dotenv==1.0.0

# Development
django-debug-toolbar==4.2.0









